+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Update the Markdown syntax to fix highlighting block quotes.
 - Support shebangs with options for syntax detection.
 - Support spaces between menu path components in Quick Actions search.
+- Update the Unicode block name list for the character inspector from Unicode 16.0.0 to Unicode 17.0.0.
 - Limit tree-sitter-based syntax highlighting to 100,000,000 characters to avoid unexpected crashes.
 - Improve overall stability.